Don't show current hour's forecast

+++ app/src/main/java/com/revengeos/weather/WeatherResponse.java
@@ -22,7 +22,7 @@ public class WeatherResponse {
     @SerializedName("clouds")
     public Clouds clouds;
     @SerializedName("dt")
-    public float dt;
+    public long dt;
     @SerializedName("id")
     public int id;
     @SerializedName("name")
